backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
sqlalchemy==2.0.40
PyJWT==2.8.0